Digital converted over to new SCPI handling

+++ platform/rp2040/platform/pi_pico_io.cpp
@@ -23,7 +23,7 @@ void CTI::PlatformIO::_statusLED(bool val) {
     gpio_put(led, val);
 }
 
-void CTI::PlatformDigital::SetOutput(int channel, bool value) {
+void CTI::PlatformDigital::SetOutput(uint16_t channel, bool value) {
     gpio_put(channel, value);
 }
